ENSWBSITES-1129: Launchbar and header update switch About ensembl and genome data & annotation in launchbar and header
