+++ src/ensembl/src/header/Header.scss
@@ -61,25 +61,15 @@
   }
 }
 
-.aboutEnsembl {
+.topRight {
   display: flex;
-  align-items: baseline;
-  margin-top: 4px;
-  font-size: 12px;
-  color: $light-blue;
-}
+  align-items: center;
+  margin-left: 20px;
 
-.logotypeAbout {
-  height: 10px;
-  fill: $light-blue;
-  margin-left: 0.5rem;
-}
-
-// For disabling temporarily the about Ensembl link in production
-.aboutEnsemblDisabled {
-  color: $medium-dark-grey;
-
-  .logotypeAbout {
-    fill: $medium-dark-grey;
+  .topRightText {
+    font-size: 13px;
+//    font-weight: $light;
+    margin-right: 25px;
+    margin-top: 2px;
   }
 }
